GoogleCloudPlateform/inversecooking/src/requirements.txt
Flask
gunicorn
numpy
scipy
matplotlib
torch==1.10.1
torchvision==0.11.2
nltk
Pillow
tqdm
lmdb
tensorflow
tensorboardX
